requirements.txt
beautifulsoup4
coverage
aiohttp
requests
pytest
asyncio-throttle
python-twitter
pandas
numpy
scikit-learn
gensim
python-dotenv
tweepy